move HighlightPhrase from Navigation to Components
