+++ .gitignore
@@ -12,7 +12,7 @@
 */out
 *.next
 *storybook-static
-*src/data/search
+*/src/data/search
 
 #logs
 npm-debug.log*
