+++ aries-site/src/components/content/index.js
@@ -4,6 +4,7 @@ export * from './ColorCompliance';
 export * from './ColorSwatch';
 export * from './ComingSoon';
 export * from './DocsMessage';
+export * from './HighlightPhrase';
 export * from './PageBackground';
 export * from './SubsectionText';
 export * from './Status';
