+++ aries-site/src/layouts/navigation/index.js
@@ -1,4 +1,3 @@
-export * from './HighlightPhrase';
 export * from './Search';
 export * from './SearchInput';
 export * from './SearchResult';
